backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
httpx==0.25.2
python-dotenv==1.0.0
celery==5.3.4
flower==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
boto3==1.34.0
pymongo==4.6.0
google-generativeai==0.3.2
openai==1.3.7
reportlab==4.0.7
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
pytesseract==0.3.10
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
streamlit==1.29.0
fastapi-mail==1.4.1
alembic==1.13.1
sqlalchemy==2.0.23
apscheduler==3.10.4
sentry-sdk[fastapi]==1.38.0
gunicorn==21.2.0
python-jose==3.3.0
cryptography==41.0.7
bcrypt==4.1.2
pyjwt==2.8.0
authlib==1.3.0
slowapi==0.1.9
limits==3.6.0
itsdangerous==2.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0
idna==3.6
jinja2==3.1.2
markupsafe==2.1.3
orjson==3.9.10
pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
